Add files via upload
This file contains both Part 1 and Part 2 solutions. I mainly toyed with lists and using simple for loops and enumerate functions to solve the questions. However, my utilization of the sum variable did trouble me to a certain extent, as I was unaware of where to shift the value of the sum draining over 1.5 hours!!! Overall Day 1 was fun, looking forward to find harder scales of questions over the coming days!